+++ src/Rx.ts
@@ -85,9 +85,7 @@ import './add/operator/skipWhile';
 import './add/operator/startWith';
 import './add/operator/subscribeOn';
 import './add/operator/switch';
-import './add/operator/exhaust';
 import './add/operator/switchMap';
-import './add/operator/exhaustMap';
 import './add/operator/switchMapTo';
 import './add/operator/take';
 import './add/operator/takeUntil';
